+++ test/runner/runner_test.dart
@@ -99,6 +99,10 @@ Usage: pub run test [files or directories...]
                                  debuggability.
                                  (defaults to on)
 
+    --[no-]retry                 Retry failing tests if configured to do so.
+                                 Useful for iterating locally on a failing test with retry configuration.
+                                 (defaults to on)
+
 ======== Output
 -r, --reporter                   The runner used to print test results.